+++ laskin/src/api/time.cpp
@@ -105,8 +105,12 @@ LASKIN_BUILTIN_WORD(w_to_number)
   const auto time = context.pop().as_time();
   number result(0.0, number::unit::second);
 
-  result += time.hour() * peelo::chrono::duration::seconds_per_hour;
-  result += time.minute() * peelo::chrono::duration::seconds_per_minute;
+  result += static_cast<double>(
+    time.hour() * peelo::chrono::duration::seconds_per_hour
+  );
+  result += static_cast<double>(
+    time.minute() * peelo::chrono::duration::seconds_per_minute
+  );
   result += time.second();
 
   context << result;
