+++ laskin/src/operator/add.cpp
@@ -130,7 +130,7 @@ namespace laskin
           U"Cannot add number to time."
         );
       }
-      delta = number_magnitude(b) * multiplier;
+      delta = static_cast<long>(number_magnitude(b) * multiplier);
     } else {
       delta = long(b);
     }
